+++ fabric/changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.19
+- Added new 1.19 sound types